+++ src/main.ts
@@ -1,5 +1,6 @@
 import { createApplication } from '@angular/platform-browser';
 import { createCustomElement } from '@angular/elements';
+import { CedarEmbeddableFieldDesignerElementComponent } from './app/element/cedar-embeddable-field-designer.element';
 import { appConfig } from './app/app.config';
 import { CedarEmbeddableDesignerElementComponent } from './app/element/cedar-embeddable-designer.element';
 import { CED_CUSTOM_ELEMENT_NAME, defineCustomElementOnce } from './app/custom-element';
@@ -20,6 +21,11 @@ import { CED_CUSTOM_ELEMENT_NAME, defineCustomElementOnce } from './app/custom-e
  */
 createApplication(appConfig)
   .then((application) => {
+    defineCustomElementOnce(
+      () => createCustomElement(CedarEmbeddableFieldDesignerElementComponent, { injector: application.injector }),
+      customElements,
+      'cedar-embeddable-field-designer',
+    );
     defineCustomElementOnce(() =>
       createCustomElement(CedarEmbeddableDesignerElementComponent, { injector: application.injector }),
     );